@@ -1,20 +1,24 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { HttpClientModule } from '@angular/common/http';
+import { CommonModule } from '@angular/common';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { ToastrModule } from 'ngx-toastr';
 
 import { AppComponent } from './app.component';
 import { CalculatorComponent } from './calculator/calculator.component';
-import { CalculatorMdbootstrapComponent } from './calculator-mdbootstrap/calculator-mdbootstrap.component';
 
 @NgModule({
    declarations: [
       AppComponent,
       CalculatorComponent,
-      CalculatorMdbootstrapComponent
    ],
    imports: [
       BrowserModule,
-      HttpClientModule
+      HttpClientModule,
+      CommonModule,
+      BrowserAnimationsModule,
+      ToastrModule.forRoot(),
    ],
    providers: [],
    bootstrap: [
